Bluetooth: Classic: SDP: correct partial response seq length
According to SDP Specification, the first parameter of
SDP_SERVICE_SEARCH_ATTR_RSP PDU is AttributeListsByteCount,
and The AttributeListsByteCount contains a count of the number
of bytes in the AttributeLists parameter.

The total length of the initial data element sequence should
indicated in the first response, even though only
a portion of this data element sequence is returned in the
first response. The remainder of this data element sequence
is returned in the second response.

Stack SDP_DATA_MTU is 200, when the total length of attrs list
larger than MTU, there have a case state.pkt_full == true,
state.att_list_size is the first partial response attrs list
length, then call repeat create_attr_list in this loop, continue
calculate att_list_size from the last attribute without copy.

Signed-off-by: zhenwei fang <fangzhenwei@bytedance.com>